Remove failing assert
This would crash if a file had been 'replaced'. I was unsure what this
was so placed this assert here. Basically can occur when a file is
removed and readded (copied from elsewhere) in the same commit, this is
often done to preserve history. Since git has automatic renaming,
there's no direct translation for this, so we can just treat it like any
old file change.