Shorten file contents lifetimes
Previously each Revision stored the contents of every touched file in
memory. This MAY get quite large either if lots of small files are
touched, or a large binary is added, or some mixture of both.

This commit changes it so that a file is only in memory if it's
currently being processed. This will likely make things slower and make
multithreading harder (if that's ever on the table).

Leaving this commit on a branch here. If OOM crashes start happening for
large commits this might be worth merging to main.