Shorten file contents lifetimes
Previously each Revision stored the contents of every touched file in
memory. This can get quite large either if lots of small files are
touched, or a large binary is added, or some mixture of both.

This commit changes it so that a file is only in memory if it's
currently being processed. We now won't spend time loading files into
memory that we're just going to ignore. This will likely make
multithreading harder (if that's ever on the table).